refactor(change_locale): rename files & add 'interface' keyword to abstract classes
