+++ lib/features/main_profile/data/datasources/remote/profile_remote_ds.dart
@@ -1,7 +1,7 @@
 import 'package:flowery_tracking_app/features/main_profile/data/models/driver_dto.dart';
 import 'package:flowery_tracking_app/features/main_profile/data/models/vehicle_dto.dart';
 
-abstract class ProfileRemoteDataSource {
+abstract interface class ProfileRemoteDataSource {
   Future<DriverDto> getProfile();
 
   Future<VehicleDto> getVehicle(String vehicleType);
