+++ lib/features/main_profile/data/datasources/remote/profile_remote_ds_impl.dart
@@ -5,7 +5,7 @@ import 'package:flowery_tracking_app/features/main_profile/data/models/vehicle_d
 import 'package:injectable/injectable.dart';
 
 @Injectable(as: ProfileRemoteDataSource)
-class ProfileRemoteDataSourceImpl extends ProfileRemoteDataSource {
+class ProfileRemoteDataSourceImpl implements ProfileRemoteDataSource {
   ApiServices apiServices;
 
   ProfileRemoteDataSourceImpl({required this.apiServices});
