+++ lib/features/main_profile/domain/usecases/profile_usecase.dart
@@ -1,22 +1,22 @@
 import 'package:flowery_tracking_app/core/errors/api_results.dart';
 import 'package:flowery_tracking_app/features/main_profile/domain/entities/driver_dto_entity.dart';
 import 'package:flowery_tracking_app/features/main_profile/domain/entities/vehicle_dto_entity.dart';
-import 'package:flowery_tracking_app/features/main_profile/domain/repositories/profile_repository_contract.dart';
+import 'package:flowery_tracking_app/features/main_profile/domain/repositories/profile_repository.dart';
 import 'package:injectable/injectable.dart';
 
 @injectable
 class ProfileUseCase {
-  ProfileRepositoryContract profileRepositoryContract;
+  ProfileRepository profileRepository;
 
-  ProfileUseCase({required this.profileRepositoryContract});
+  ProfileUseCase({required this.profileRepository});
 
   Future<ApiResult<DriverDtoEntity>> getProfile() async {
-    var driverResult = await profileRepositoryContract.getProfile();
+    var driverResult = await profileRepository.getProfile();
     return driverResult;
   }
 
   Future<ApiResult<VehicleDtoEntity>> getVehicle(String vehicleType) async {
-    var vehicleResult = await profileRepositoryContract.getVehicle(vehicleType);
+    var vehicleResult = await profileRepository.getVehicle(vehicleType);
     return vehicleResult;
   }
 }
